+++ halotools/mock_observables/void_statistics/__init__.py
@@ -1,0 +1,6 @@
+""" Sub-package containing functions used to calculate void probability functions. 
+"""
+from __future__ import absolute_import 
+
+from .void_prob_func import void_prob_func
+from .underdensity_prob_func import underdensity_prob_func
